fund-analyser-ml/requirements.txt
bdateutil==0.1
beautifulsoup4==4.6.0
keras==2.0.6
lxml==3.8.0
matplotlib==2.0.2
numpy==1.13.1
pandas==0.20.2
requests==2.18.1
scipy==0.19.1
scikit-learn==0.18.2
theano==0.9.0